+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="x.x.x"></a>
+# x.x.x (xxx-xx-xx)
+
+
+### Bug Fixes
+
+* **69:** Respect of all parents classes and their methods in mocks
+
+
 <a name="9.1.0"></a>
 # 9.1.0 (2020-03-21)